requirements.txt
streamlit
pandas
matplotlib
openai
fpdf
sendgrid
Pillow
